wasm_interp: simplified ImportDispatcher
We don't really need a trait for the 'import module' concept.
